add 1 second
add test

docker

+++ src/main/java/ua/com/alevel/plannerbox/service/impl/UserServiceImpl.java
@@ -1,35 +1,42 @@
 package ua.com.alevel.plannerbox.service.impl;
 
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ua.com.alevel.plannerbox.dto.UserDto;
 import ua.com.alevel.plannerbox.entity.User;
-import ua.com.alevel.plannerbox.entity.UserRole;
 import ua.com.alevel.plannerbox.entity.status.UserStatus;
 import ua.com.alevel.plannerbox.exceptions.UserAlreadyExistException;
+import ua.com.alevel.plannerbox.exceptions.UserNotFoundException;
+import ua.com.alevel.plannerbox.mapper.UserMapper;
+import ua.com.alevel.plannerbox.repository.RoleRepository;
 import ua.com.alevel.plannerbox.repository.UserRepository;
-import ua.com.alevel.plannerbox.service.UserRoleService;
+import ua.com.alevel.plannerbox.security.SecurityContextHelper;
 import ua.com.alevel.plannerbox.service.UserService;
 
 import java.time.LocalDateTime;
-import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
+
 
 @Service
 @Slf4j
 @RequiredArgsConstructor
-@Transactional
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
-    private final UserRoleService userRoleService;
+    private final RoleRepository roleRepository;
     private final PasswordEncoder passwordEncoder;
+    private final SecurityContextHelper securityContextHelper;
+    private final UserMapper userMapper;
 
     @Override
+    @Transactional
     public User createAdmin() {
         User admin = new User();
         admin.setUsername("admin");
@@ -39,26 +46,26 @@ public class UserServiceImpl implements UserService {
         admin.setPassword(passwordEncoder.encode("adminPlanner"));
         admin.setStatus(UserStatus.ACTIVE);
         admin.setCreated(LocalDateTime.now());
+        admin.setUpdated(LocalDateTime.now());
         log.info("Admin - {} successfully registered", admin.getUsername());
         return userRepository.save(admin);
     }
 
-    // TODO add role "USER" for default
     @Override
-    public User registerUser(User user) {
-        user.setUsername(user.getUsername());
-        user.setFirstName(user.getFirstName());
-        user.setLastName(user.getLastName());
-        user.setEmail(user.getEmail());
-        user.setPassword(passwordEncoder.encode(user.getPassword()));
+    @Transactional
+    public User registerUser(UserDto userDto) {
+        if (userRepository.findByUsername(userDto.getUsername()) != null) {
+            throw new UserAlreadyExistException("Username already exist");
+        }
+        User user = userMapper.userFromDto(userDto);
+        user.setPassword(passwordEncoder.encode(userDto.getPassword()));
         user.setStatus(UserStatus.ACTIVE);
         user.setCreated(LocalDateTime.now());
-        log.info("User: {} successfully registered", user.getUsername());
-        return userRepository.save(user);
-    }
-
-    public Set<UserRole> addDefaultUserRole() {
-        return Collections.singleton(userRoleService.findByRole("USER"));
+        user.setUpdated(LocalDateTime.now());
+        user.setRoles(Set.of(roleRepository.findByRole("USER")));
+        User registeredConsumer = userRepository.save(user);
+        log.info("User: {} successfully registered", registeredConsumer.getUsername());
+        return registeredConsumer;
     }
 
     @Override
@@ -68,32 +75,44 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public List<User> getAllUser() {
+    @Transactional
+    public List<User> findAllUsers() {
+        List<User> allUsers = userRepository.findAll();
         log.info("Fetching all users");
-        return userRepository.findAll();
+        return allUsers;
     }
 
     @Override
-    public User findUserById(Long id) {
-        User result = userRepository.findById(id).orElse(null);
-        if (result == null) {
+    @Transactional
+    public Optional<User> findUserById(Long id) {
+        log.info("Start find user by id: {}", id);
+        Optional<User> user = userRepository.findUserById(id);
+        if (user.isEmpty()) {
             log.warn("User with id {} does not exist", id);
-            throw new UsernameNotFoundException("User with id " + id + " does not exist");
         }
-        log.info("Fetching user by id: {}", result);
-        return result;
+        return user;
     }
 
     @Override
-    public void deleteUser(Long id) {
+    @Transactional
+    public void deleteUser(Long id) throws UserNotFoundException {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException("User with id " + id + " does not found");
+        }
+        log.info("Start delete user by id: {}", id);
         userRepository.deleteById(id);
         log.info("User with id: {} successfully delete", id);
     }
 
-    private void userNameValidity(User user) throws UserAlreadyExistException {
-        User consumerName = userRepository.findByUsername(user.getUsername());
-        if (getAllUser().contains(consumerName)) {
-            throw new UserAlreadyExistException("username is already taken"); // TODO or log.warn???
-        }
+    @Override
+    @Transactional
+    public User updateUser(UserDto userDto) {
+        String currentUser = securityContextHelper.getCurrentUser().getUsername();
+        User user = userRepository.findByUsernameAndId(currentUser, userDto.getId());
+        user.setUpdated(LocalDateTime.now());
+        user.setPassword(passwordEncoder.encode(userDto.getPassword()));
+        userMapper.updateUser(userDto, user);
+        log.info("User with username {} and user id: {} successfully updated", currentUser, userDto.getId());
+        return userRepository.save(user);
     }
 }
